requirements.txt
    Django>=4.2,<5.0
    djangorestframework>=3.14
    requests>=2.28